+++ axelrod/tests/unit/test_filters.py
@@ -1,6 +1,6 @@
 import operator
 import unittest
-from hypothesis import given, example
+from hypothesis import given, example, settings
 from hypothesis.strategies import integers
 from axelrod.strategies._filters import *
 from axelrod import filtered_strategies
@@ -35,6 +35,7 @@ class TestFilters(unittest.TestCase):
         larger=integers(min_value=11, max_value=100),
     )
     @example(smaller=0, larger=float('inf'))
+    @settings(max_examples=5, max_iterations=20)
     def test_inequality_filter(self, smaller, larger):
         self.assertTrue(passes_operator_filter(
             self.TestStrategy, 'memory_depth', smaller, operator.ge))
@@ -60,6 +61,7 @@ class TestFilters(unittest.TestCase):
         larger=integers(min_value=11, max_value=100),
     )
     @example(smaller=0, larger=float('inf'))
+    @settings(max_examples=5, max_iterations=20)
     def test_passes_filterset(self, smaller, larger):
 
         full_passing_filterset_1 = {